+++ Acr.Ble.Server.Android/GattCharacteristic.cs
@@ -11,7 +11,7 @@ using Observable = System.Reactive.Linq.Observable;
 
 namespace Acr.Ble.Server
 {
-    public class GattCharacteristic : AbstractGattCharacteristic
+    public class GattCharacteristic : AbstractGattCharacteristic, IDroidGattCharacteristic
     {
         public static readonly Guid NotifyDescriptorId = new Guid("00002902-0000-1000-8000-00805f9b34fb");
         public static readonly UUID NotifyDescriptorUuid = UUID.FromString("00002902-0000-1000-8000-00805f9b34fb");