+++ Acr.Ble.Server.Android/IDroidGattCharacteristic.cs
@@ -1,0 +1,11 @@
+using System;
+using Android.Bluetooth;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IDroidGattCharacteristic : IGattCharacteristic
+    {
+        BluetoothGattCharacteristic Native { get; }
+    }
+}