+++ Acr.Ble.Server.Android/IDroidGattDescriptor.cs
@@ -1,0 +1,11 @@
+using System;
+using Android.Bluetooth;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IDroidGattDescriptor : IGattDescriptor
+    {
+        BluetoothGattDescriptor Native { get; }
+    }
+}