+++ Acr.Ble.Server.Android/IDroidGattService.cs
@@ -1,0 +1,11 @@
+using System;
+using Android.Bluetooth;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IDroidGattService : IGattService
+    {
+        BluetoothGattService Native { get; }
+    }
+}