+++ Acr.Ble.Server.Uwp/IUwpGattService.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Threading.Tasks;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IUwpGattService : IGattService
+    {
+        Task Init();
+        void Stop();
+    }
+}