+++ Acr.Ble.Server.iOS/IIosGattCharacteristic.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using CoreBluetooth;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IIosGattCharacteristic : IGattCharacteristic
+    {
+        CBMutableCharacteristic Native { get; }
+    }
+}