+++ Acr.Ble.Server.iOS/IIosGattDescriptor.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using CoreBluetooth;
+
+
+namespace Acr.Ble.Server
+{
+    public interface IIosGattDescriptor : IGattDescriptor
+    {
+        CBMutableDescriptor Native { get; }
+    }
+}